* binn codec refactoring, * binn codec armv5 alignment compatilility

+++ src/jbl/binn.h
@@ -84,8 +84,8 @@ typedef int BOOL;
 typedef __int64 int64;
 typedef unsigned __int64 uint64;
 #else
-typedef long long int int64;
-typedef unsigned long long int uint64;
+typedef int64_t int64;
+typedef uint64_t uint64;
 #endif
 #endif
